+++ packages/core-react-ui/src/badges/badge.treat.ts
@@ -15,10 +15,13 @@ export const base = style(({ theme, cssTheme }: Theme) => ({
   backgroundColor: '#555555',
   textTransform: 'lowercase',
   fontSize: 12,
+  lineHeight: 1.25,
   fontWeight: 400,
   fontFamily: 'Verdana,Geneva,DejaVu Sans,sans-serif',
   textShadow: '0px .1em 0px rgba(0, 0, 0, 0.5)',
   color: 'white',
+
+  background: 'linear-gradient(180deg, rgba(187, 187, 187, .1) 0%, rgba(0, 0, 0, .1) 100%)',
 }));
 
 export const name = style(({ theme, cssTheme }: Theme) => ({
@@ -29,7 +32,7 @@ export const name = style(({ theme, cssTheme }: Theme) => ({
 export const nameSpan = style(({ theme, cssTheme }: Theme) => ({
   display: 'inline-block',
   padding: '.35em .5em',
-  background: 'linear-gradient(180deg, rgba(187, 187, 187, .1) 0%, rgba(0, 0, 0, .1) 100%)',
+  backgroundColor: `rgb(85, 85, 85)`,
 }));
 
 export const status = style(({ theme, cssTheme }: Theme) => ({
@@ -41,23 +44,6 @@ export const status = style(({ theme, cssTheme }: Theme) => ({
 export const statusSpan = style(({ theme, cssTheme }: Theme) => ({
   padding: '.35em .5em',
   display: 'inline-block',
-  background: 'linear-gradient(180deg, rgba(187, 187, 187, .1) 0%, rgba(0, 0, 0, .1) 100%)',
-  // background: getCSSLinearGradient({
-  //   angle: {
-  //     value: 0,
-  //     unit: ANGLE_UNIT.DEG,
-  //   },
-  //   stops: [
-  //     {
-  //       color: theme.colors.colors.grey[900],
-  //       position: 0,
-  //     },
-  //     {
-  //       color: theme.colors.colors.grey[100],
-  //       position: 100,
-  //     },
-  //   ],
-  // }),
 }));
 
 /**
@@ -65,3 +51,15 @@ export const statusSpan = style(({ theme, cssTheme }: Theme) => ({
  * Variants
  *
  */
+
+export const neutral = style(({ theme, cssTheme }: Theme) => ({
+  backgroundColor: '#a7a7a7',
+}));
+
+export const success = style(({ theme, cssTheme }: Theme) => ({
+  backgroundColor: cssTheme.colors.colorIntents.successAction,
+}));
+
+export const error = style(({ theme, cssTheme }: Theme) => ({
+  backgroundColor: cssTheme.colors.colorIntents.dangerAction,
+}));