+++ packages/core-react-ui/src/primitive/primitive.css.ts
@@ -1,0 +1,5 @@
+import { style } from '@vanilla-extract/css';
+
+export const reset = style({
+  lineHeight: 1,
+});